rename controller Smart -> Adaptive
